feat: Add ExperimentWhitelistService (#129)
Summary:

Add ExperimentWhitelistService, which returns decisions based on the contents of forcedVariations in experiments in the datafile. NewCompositeExperimentService creates an ExperimentWhitelistService as its first child service, taking precedence over normal bucketing.

To support ExperimentWhitelistService, I added a Whitelist field to the Experiment entity. The value comes directly from forcedVariations of the datafile experiment. I named it Whitelist to match the term we use to talk about this kind of override, even though it's not named that in the datafile.

Test plan:

Added unit tests