fix(ExperimentWhitelistService): Fix finding variation by key (#155)
Previously, GetDecision of ExperimentWhitelistService was incorrectly using variation key to look up variations in the Variations map of an Experiment struct, in which keys are variation IDs, not variation keys. 

With this fix, we iterate over variations to find the one with the desired key. To avoid needing to search all variations, we can add a map keyed by variation key as a follow up.